requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
auth==1.3.0
blinker==1.9.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
Flask==3.1.3
flask-cors==6.0.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pip-tools==7.5.3
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
waitress==3.0.2
Werkzeug==3.1.6
wheel==0.46.3
